added more capability checks and more logging

+++ app/Http/Controllers/Api/LessonController.php
@@ -91,22 +91,41 @@ class LessonController extends SMController
 
     public function store()
     {
-        $prev_lesson =  Lesson::whereSiteId($this->site->id)->where('module_id','!=','0')->orderBy('sort_order' , 'desc')->first();
-        if($prev_lesson)
-            Input::merge(array('sort_order'=>$prev_lesson->sort_order + 1));
-        else
-            Input::merge(array('sort_order'=>1));
+		if( \SMRole::userHasAccess( $this->site->id, 'manage_content', \Auth::user()->id ) )
+		{
+			\Input::merge(['site_id' => $this->site->id ]);
+			$prev_lesson = Lesson::whereSiteId( $this->site->id )->where( 'module_id', '!=', '0' )
+				->orderBy( 'sort_order', 'desc' )->first();
+			if( $prev_lesson )
+				Input::merge( array( 'sort_order' => $prev_lesson->sort_order + 1 ) );
+			else
+				Input::merge( array( 'sort_order' => 1 ) );
 
-        $stored = parent::store();
-        $courseTitle = SiteMetaData::whereSiteId($this->site->id)->whereKey('course_title')->first();
-        
-        if($courseTitle!=null)
-            $notification=array_merge(array('title' => 'lesson alert'),array('content' => $stored->title),array('start_date' => Carbon::now()),array('end_date' => Carbon::now()->addDay()),array('site_id' => $this->site->id),array('type' => 'lesson'));
-        else
-            $notification=array_merge(array('title' => 'lesson alert'),array('start_date' => Carbon::now()),array('end_date' => Carbon::now()->addDay()),array('content' => $stored->title),array('site_id' => $this->site->id),array('type' => 'lesson'));
-        if($stored->access_level_type!=4)
-            SiteNotice::create($notification);
-        return $stored;
+			$stored      = parent::store();
+			$courseTitle = SiteMetaData::whereSiteId( $this->site->id )->whereKey( 'course_title' )->first();
+
+			if( $courseTitle != null )
+				$notification = array_merge( array( 'title' => 'lesson alert' ), array( 'content' => $stored->title ), array( 'start_date' => Carbon::now() ), array( 'end_date' => Carbon::now()
+					->addDay() ), array( 'site_id' => $this->site->id ), array( 'type' => 'lesson' ) );
+			else
+				$notification = array_merge( array( 'title' => 'lesson alert' ), array( 'start_date' => Carbon::now() ), array( 'end_date' => Carbon::now()
+					->addDay() ), array( 'content' => $stored->title ), array( 'site_id' => $this->site->id ), array( 'type' => 'lesson' ) );
+			if( $stored->access_level_type != 4 )
+				SiteNotice::create( $notification );
+
+			\App\Models\Event::Log( 'lesson-created', array(
+				'site_id' => $this->site->id,
+				'user_id' => \Auth::user()->id,
+				'lesson-title' => $stored->title,
+				'lesson-id' => $stored->id
+			) );
+
+			return $stored;
+		}
+		else
+		{
+			return [];
+		}
     }
 
     public function update($model)
@@ -114,29 +133,63 @@ class LessonController extends SMController
         $newpermalink=\Input::get("permalink");
         if(!\Input::get('site_id'))
             \Input::merge(array('site_id' => $this->site->id ));
-        
-        //homepageURLupdate
-        SiteMetaData::where("key","homepage_url")->where("site_id",$this->site->id)->where("value","lesson/".$model->permalink)->update(array("value" => "lesson/".\Input::get("permalink") ));
-        SiteFooterMenuItem::where("url","lesson/".$model->permalink)->where("site_id",$this->site->id)->update(array("url"=>"lesson/".\Input::get("permalink")));
-        SiteMenuItem::where("url","lesson/".$model->permalink)->where("site_id",$this->site->id)->update(array("url"=>"lesson/".\Input::get("permalink")));
-        
-        return $model->update(\Input::except('_method' , 'access'));
+
+		if( \SMRole::userHasAccess( $this->site->id, 'manage_content', \Auth::user()->id ) )
+		{
+			//homepageURLupdate
+			SiteMetaData::where( "key", "homepage_url" )->where( "site_id", $this->site->id )
+				->where( "value", "lesson/" . $model->permalink )
+				->update( array( "value" => "lesson/" . \Input::get( "permalink" ) ) );
+			SiteFooterMenuItem::where( "url", "lesson/" . $model->permalink )->where( "site_id", $this->site->id )
+				->update( array( "url" => "lesson/" . \Input::get( "permalink" ) ) );
+			SiteMenuItem::where( "url", "lesson/" . $model->permalink )->where( "site_id", $this->site->id )
+				->update( array( "url" => "lesson/" . \Input::get( "permalink" ) ) );
+
+			\App\Models\Event::Log( 'lesson-updated', array(
+				'site_id' => $this->site->id,
+				'user_id' => \Auth::user()->id,
+				'lesson-title' => $model->title,
+				'lesson-id' => $model->id
+			) );
+
+			return $model->update( \Input::except( '_method', 'access' ) );
+		}
+		else
+		{
+			return $model;
+		}
     }
 
-    public function destroy($model){
-        $notification = SiteNotice::whereContent($model->title)->first();
-        if($notification!=null)
-            $notification->delete();
-        else
-            \Log::info('koi msla he vai :'.$model->title);
-        $model->site->total_lessons = $model->site->total_lessons - 1;
-        $model->site->save();
+    public function destroy($model)
+	{
+		if( \SMRole::userHasAccess( $this->site->id, 'manage_content', \Auth::user()->id ) )
+		{
+			$notification = SiteNotice::whereContent( $model->title )->first();
+			if( $notification != null )
+				$notification->delete();
+			else
+				\Log::info( 'koi msla he vai :' . $model->title );
+			$model->site->total_lessons = $model->site->total_lessons - 1;
+			$model->site->save();
 
-		$permalinks = Permalink::whereSiteId($model->site_id)->whereTargetId($model->id)->whereType($model->getTable())->get();
-		foreach( $permalinks as $permalink )
-			$permalink->delete();
+			$permalinks = Permalink::whereSiteId( $model->site_id )->whereTargetId( $model->id )
+				->whereType( $model->getTable() )->get();
+			foreach( $permalinks as $permalink )
+				$permalink->delete();
 
-        return parent::destroy($model);
+			\App\Models\Event::Log( 'lesson-deleted', array(
+				'site_id' => $this->site->id,
+				'user_id' => \Auth::user()->id,
+				'lesson-title' => $model->title,
+				'lesson-id' => $model->id
+			) );
+
+			return parent::destroy( $model );
+		}
+		else
+		{
+			return [];
+		}
     }
 
     public function addAllVideos(){
